Refactor: StyledText 리팩토링
